Rooms, traversable floors, and walls.
Most of the classes are comprised of a simple rectangle. The room class is comprised of a wall which is a rectangle. A room also has a floor, but a floor is considered traversable unlike a wall. All the information about each class is stored in a class called buildingBlocks to allow them to be easily written into a text file.