My final contribution (for the competition).
The competition is almost over and we're just going to submit what we have. I only added some methods in room, walls, and doors. They were an attempt to find an error that we resolved later. The LocationReader class should be done, but we never got around to implementing it. I plan on doing that when I am not so tired.